+++ OpenSim/Simulation/InverseKinematicsSolver.cpp
@@ -89,6 +89,37 @@ InverseKinematicsSolver::InverseKinematicsSolver(const Model& model,
             log_warn("WARNING: InverseKinematicsSolver found only {} markers to track.", cnt);
 
     }
+    if (_orientationsReference && _orientationsReference->getNumRefs() > 0) {
+        const SimTK::Array_<std::string>& sensorNames =
+                _orientationsReference
+                        ->getNames(); // size and content as in orientations file
+
+        if (sensorNames.size() < 1) {
+            log_error("InverseKinematicsSolver: No orientation data is available from provided "
+                      "source.");
+            throw Exception("InverseKinematicsSolver: No orientation data is available "
+                            "from the provided source.");
+        }
+        int cnt = 0;
+        const auto onFrames = getModel().getComponentList<PhysicalFrame>();
+
+        for (const auto& modelFrame : onFrames) {
+            const std::string& modelFrameName = modelFrame.getName();
+            auto found = std::find(
+                    sensorNames.begin(), sensorNames.end(), modelFrameName);
+            if (found != sensorNames.end()) { cnt++; }
+        }
+
+        if (cnt < 1) {
+            log_error("InverseKinematicsSolver: Orientation data does not "
+                      "correspond to any model frames.");
+            throw Exception("InverseKinematicsSolver: Orientation data does not "
+                            "correspond to any model frames.");
+        }
+        log_info("InverseKinematicsSolver found {} model frames to track.",
+                    cnt);
+
+    }
 }
 
 int InverseKinematicsSolver::getNumMarkersInUse() const
@@ -307,7 +338,7 @@ double InverseKinematicsSolver::
 double InverseKinematicsSolver::computeCurrentOrientationError(int osensorIndex)
 {
     if (osensorIndex >= 0 && 
-        osensorIndex < _markerAssemblyCondition->getNumMarkers()) {
+        osensorIndex < _orientationAssemblyCondition->getNumOSensors()) {
         return _orientationAssemblyCondition->
             findCurrentOSensorError(
                 SimTK::OrientationSensors::OSensorIx(osensorIndex));
@@ -315,7 +346,7 @@ double InverseKinematicsSolver::computeCurrentOrientationError(int osensorIndex)
     else
         throw Exception(
             "InverseKinematicsSolver::computeCurrentOrientationError: "
-            "invalid markerIndex." );
+            "invalid sensor Index." );
 }
 
 /* Compute and return the distance between all model o-sensors and their observations. */